chore: update frontend to reflect backend book restructure

+++ src/getters.rs
@@ -30,7 +30,7 @@ async fn get_movie(data: Data<AppState>, id: Path<u64>) -> impl Responder {
 }
 
 #[get("/api/book/{id}")]
-async fn get_book(data: Data<AppState>, id: Path<u64>) -> impl Responder {
+async fn get_book(data: Data<AppState>, id: Path<u32>) -> impl Responder {
     match data.0.lock().await.books.get(&id) {
         Some(book) => HttpResponse::Ok().json(book),
         None => HttpResponse::NotFound().finish(),
